Feat: keep mapping same DTO for bot and dashboard
